Implement CustomRepositoryHelper for managing custom repositories in NSI handler
